requirements.txt
aiofiles==24.1.0
alembic==1.16.1
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.9.0
attrs==25.3.0
Automat==25.4.16
backoff==2.2.1
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
coloredlogs==14.0
constantly==23.10.4
cryptography==45.0.4
cssselect==1.3.0
dagster==1.10.19
dagster-pipes==1.10.19
dagster_shared==1.10.19
dataclasses-json==0.6.7
defusedxml==0.7.1
docstring_parser==0.16
emoji==2.14.1
exceptiongroup==1.3.0
feedparser==6.0.11
filelock==3.18.0
filetype==1.2.0
fsspec==2025.5.1
greenlet==3.2.3
grpcio==1.73.0
grpcio-health-checking==1.71.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
itemadapter==0.11.0
itemloaders==1.3.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
langdetect==1.0.9
lxml==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.6
olefile==0.47
packaging==25.0
parsel==1.10.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
pluggy==1.6.0
Protego==0.4.0
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyDispatcher==2.0.7
Pygments==2.19.1
pyOpenSSL==25.1.0
pypdf==5.6.0
pypdfium2==4.30.1
pytest==8.4.0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.2
queuelib==1.8.0
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==14.0.0
Scrapy==2.13.2
service-identity==24.2.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
structlog==25.4.0
tabulate==0.9.0
tldextract==5.3.0
tomli==2.2.1
tomlkit==0.13.3
toposort==1.10
tqdm==4.67.1
Twisted==25.5.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
universal_pathlib==0.2.6
unstructured==0.17.2
unstructured-client==0.36.0
urllib3==2.4.0
w3lib==2.3.1
watchdog==5.0.3
webencodings==0.5.1
wrapt==1.17.2
zope.interface==7.2
